Refactored the path aliases.
