+++ .storybook/webpack.config.js
@@ -19,8 +19,7 @@ module.exports = ({ config }) => {
   config.performance = { hints: false };
   config.resolve.alias = {
     ...config.resolve.alias,
-    '~': path.resolve(__dirname, '../src/app/'),
-    '~t': path.resolve(__dirname, '../src/__tests__/')
+    '~': path.resolve(__dirname, '../src/')
   };
   config.resolve.extensions.push('.ts', '.tsx');
