+++ src/__tests__/App.test.tsx
@@ -1,6 +1,6 @@
 import Enzyme from 'enzyme';
 import React from 'react';
-import App from '~/App';
+import App from '~/app/App';
 
 describe('<App />', () => {
   beforeEach(() => {
