+++ src/app/Dummy.stories.tsx
@@ -1,6 +1,6 @@
 import { storiesOf } from '@storybook/react';
 import React from 'react';
-import Dummy from '~/Dummy';
+import Dummy from '~/app/Dummy';
 import { IInteractions } from './stories.toc';
 
 /** Interactions definition for BackstopJS. */
