+++ tsconfig.json
@@ -9,8 +9,7 @@
     "moduleResolution": "node",
     "outDir": "./dist",
     "paths": {
-      "~t*": ["src/__tests__*"],
-      "~*": ["src/app*"]
+      "~*": ["src*"]
     },
     "resolveJsonModule": true,
     "sourceMap": true,
